go111 static sample: fix app.yaml
Hi,

In the main.go file, `/public` is used for the http handler, and the indexData (passed to the template). Shouldn't the yaml file also use `/public` for when served by appengine?

It probably still works as is anyway, since `/public` won't match any route from `app.yaml`, and the request will be redirected to the go binary, which will handle it. But it seems the idea of this example was to have both (there are comments in the app.yaml and main.go suggesting that the static/public handlers are handled by appengine in production).

Alternatively, I change main.go to use /static...